% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpg_annotate.R
\name{annotate_and_save}
\alias{annotate_and_save}
\title{CpG annotate and save csv files}
\usage{
annotate_and_save(file_names, input_folder, output_folder)
}
\arguments{
\item{file_names}{list of csv file names that need to be annotated}

\item{input_folder}{location of the EWAS results in csv format}

\item{output_folder}{path to save the output files}
}
\value{
this function will return an R data object as a result of the CpG wide association using cpg.assoc function
}
\description{
CpG annotate and save csv files
}
