% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwas_assign_KEGG_super_pathways.R
\name{assignGroup}
\alias{assignGroup}
\title{Assigns KEGG super pathways to the mummichog results (load curated list first)}
\usage{
assignGroup(data, group_column, pathway_column, pathway_list)
}
\arguments{
\item{data}{input dataframe that contain all mummichog results}

\item{group_column}{dummy variable that takes the group names}

\item{pathway_column}{variable that has the pathway names}

\item{pathway_list}{load this from the separate R file available in this library}
}
\description{
Assigns KEGG super pathways to the mummichog results (load curated list first)
}
