% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkmr_sex_spec.R
\name{bkmr_sex}
\alias{bkmr_sex}
\title{BKMR wrapper performing sex specific loops. Load Matrix and bindrcpp packages}
\usage{
bkmr_sex(
  data,
  folder_path,
  include_sex = TRUE,
  ln_mixture_vars = NULL,
  outcome_vars = NULL,
  covariate_vars = NULL,
  groups,
  iter = iter
)
}
\arguments{
\item{data}{input of a dataframe that contain all the variables mentioned in the wrapper}

\item{folder_path}{input folder location to save the output in RDA format}

\item{include_sex}{input TRUE or FASE to perform sex stratified analysis}

\item{ln_mixture_vars}{input list of chemicals that need to be considered as mixture}

\item{outcome_vars}{input list of outcome variables}

\item{covariate_vars}{input list of covariates}

\item{groups}{input grouping structure of the chemicals considered as mixture, these values should be integers}

\item{iter}{input a numeric value for number of iterations}
}
\value{
this function will return output from bkmr model in RDA format
}
\description{
BKMR wrapper performing sex specific loops. Load Matrix and bindrcpp packages
}
