% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwas_find_p_cutoff_manually.R
\name{calculatePvalCutoff}
\alias{calculatePvalCutoff}
\title{Manually get p-value cutoffs based on the 500 features with lowest p-values
use along with the runMummichog function}
\usage{
calculatePvalCutoff(input_file, default_cutoff)
}
\description{
Manually get p-value cutoffs based on the 500 features with lowest p-values
use along with the runMummichog function
}
