% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwas_combine_mummichog_results.R
\name{combine_csv_files}
\alias{combine_csv_files}
\title{Combines mummuchog output files into a single dataframe}
\usage{
combine_csv_files(folder_location)
}
\arguments{
\item{folder_location}{Specify folder location that contain all the mummichog results}
}
\value{
returns a dataframe with all results into the global environment
}
\description{
Combines mummuchog output files into a single dataframe
}
\examples{
\donttest{
combine_csv_files("~/Documents/phth_phe_MWAS/result/enrich_imp")
}


}
