% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas_enrichment_summary.R
\name{enrichment_summary}
\alias{enrichment_summary}
\title{ewas enrichemnt summary}
\usage{
enrichment_summary(folder_path, output_folder)
}
\arguments{
\item{folder_path}{folder location that contain the enriched results in csv format}

\item{output_folder}{folder to save results}
}
\value{
returns summary of all ewas enrichment results
}
\description{
ewas enrichemnt summary
}
