% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas_loop_pll.R
\name{ewas_loop_pll}
\alias{ewas_loop_pll}
\title{Performs EWAS using CpGassoc library. Allows flexibility to loop for multiple exposure variables and methylation data from multiple tissues.}
\usage{
ewas_loop_pll(
  exposures,
  covariates,
  bVals_list,
  bVals_names,
  dat_pheno,
  outputFolder
)
}
\arguments{
\item{exposures}{list of exposure variables}

\item{covariates}{list of covariates to adjust for associations}

\item{bVals_list}{processed beta values set from illumina methylation}

\item{bVals_names}{name of the beta value datasets - will be used for naming the output files}

\item{dat_pheno}{dataframe that contain exposures, covariates and cell type counts}

\item{outputFolder}{path to save the output files}
}
\value{
this function will return an R data object as a result of the CpG wide association using cpg.assoc function
}
\description{
Performs EWAS using CpGassoc library. Allows flexibility to loop for multiple exposure variables and methylation data from multiple tissues.
}
