% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas_summary.R
\name{ewas_summary}
\alias{ewas_summary}
\title{Summarize results from EWAS analysis}
\usage{
ewas_summary(folder_path)
}
\arguments{
\item{folder_path}{provide location of a folder that contains EWAS results in csv format}
}
\value{
returns a summary of the associations by exposure-methylation data combinations
}
\description{
Summarize results from EWAS analysis
}
