% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkmr_result_extract.R
\name{extract_data_from_folder}
\alias{extract_data_from_folder}
\title{Extract results from the bkmr RDA files}
\usage{
extract_data_from_folder(folder_location, select_iterations, estimate_sequence)
}
\arguments{
\item{folder_location}{input folder location that contain bkmr output in RDA format}

\item{select_iterations}{select iterations among the total number of iterations}

\item{estimate_sequence}{input sequence and intervals to display overall estimates.}
}
\value{
this function will return a list of two objects the first is PIPs and second is overall risk estimates
}
\description{
Extract results from the bkmr RDA files
}
