% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkmr_pips_forShiny.R
\name{extract_pips}
\alias{extract_pips}
\title{Extract bkmr PIPs from all RDA files - prep for Shiny object}
\usage{
extract_pips(folder_location)
}
\arguments{
\item{folder_location}{input a folder location that contain bkmr output in RDA format}
}
\value{
This function will return a dataframe that contain information from file names and PIPs
}
\description{
Extract bkmr PIPs from all RDA files - prep for Shiny object
}
