% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgcomp_weight.R
\name{extract_weights}
\alias{extract_weights}
\title{Extract chemical weights contributed to the joint association using non bootstrap quantile g-computation.}
\usage{
extract_weights(folder_location)
}
\arguments{
\item{folder_location}{input folder location that contain the RDA output files from the non bootstrap quantile g computation. this function will extract from files that contain nb in the filename.}
}
\value{
returns a dataframe that will contain the RDA fileaname, chemical name and corresponding weight
}
\description{
Extract chemical weights contributed to the joint association using non bootstrap quantile g-computation.
}
