% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgcomp_res_extract.R
\name{get_gcomp_estimates}
\alias{get_gcomp_estimates}
\title{Extract results from the qgcomp output RDA files. Iterates though the RDA files and compiles results in tibble format}
\usage{
get_gcomp_estimates(results_folder)
}
\arguments{
\item{results_folder}{input a folder location that contain qgcomp RDA output files}
}
\value{
results a dataframe that will contain bootstap or not, gender, cohort, outcome, estimate, confidence intervals
}
\description{
Extract results from the qgcomp output RDA files. Iterates though the RDA files and compiles results in tibble format
}
