% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_linear_reg.R
\name{lm_func}
\alias{lm_func}
\title{Multiple linear regression wrapper function that will take a list of exposure and outcome variables. Returns a matrix of results between each combination}
\usage{
lm_func(
  dependent_vars,
  independent_vars,
  data,
  covariates,
  include_sex = TRUE,
  include_cohort = TRUE,
  conf_level = 0.95
)
}
\arguments{
\item{dependent_vars}{input a list of outcome variables}

\item{independent_vars}{input a list of exposure variables}

\item{data}{input a dataset tested with dataframe and tibble formats}

\item{covariates}{input a list of confounders}

\item{include_sex}{input either TRUE or FALSE to perform stratified analysis}

\item{include_cohort}{input either TRUE or FALSE to perform stratified analysis}

\item{conf_level}{input confidence interval level. Default to yield 95 percent Confidence interval}
}
\value{
This function will return a dataframe that will contain exposure, outcome, cohort level, sex level, regression coefficient, 95 percent confidence interval
}
\description{
Multiple linear regression wrapper function that will take a list of exposure and outcome variables. Returns a matrix of results between each combination
}
