% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwas_seq_match.R
\name{match_sequence}
\alias{match_sequence}
\title{match sequence - validation step before getting into MWAS loop
match the exposure dataset with the metabolome sequence
metabolome each observation is a feature and each variable is a subject
Whereas the exposure/covariate dataset each observation is a subject}
\usage{
match_sequence(df1, df2, variable_name)
}
\arguments{
\item{df1}{A dataframe to match}

\item{df2}{second dataframe to match with the df1}
}
\value{
print's sequence of identifiers that matched between datasets
}
\description{
match sequence - validation step before getting into MWAS loop
match the exposure dataset with the metabolome sequence
metabolome each observation is a feature and each variable is a subject
Whereas the exposure/covariate dataset each observation is a subject
}
\examples{
\donttest{
match_sequence(df1= mom_dat, #dataset with exposure/outcome variables
               df2=mom_ft, #dataset with metabolome features
               variable_name="participant_id_mom")
}

}
