% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewas_gene_enrichment.R
\name{perform_gene_enrichment}
\alias{perform_gene_enrichment}
\title{Gene enrichment for EWAS files}
\usage{
perform_gene_enrichment(input_folder, output_folder)
}
\arguments{
\item{input_folder}{Annotated version of EWAS results in csv format}

\item{output_folder}{location to save the results}
}
\value{
returns a csv that will contain pathways associated with input CpGs with q-value <0.05 as threshold
}
\description{
Gene enrichment for EWAS files
}
