% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_env.R
\name{print_version_info}
\alias{print_version_info}
\title{Prints detailed summary of the R environment including libraries and dependencies}
\usage{
print_version_info()
}
\value{
displays R environment summary
}
\description{
Prints detailed summary of the R environment including libraries and dependencies
}
