% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_methylation_data.R
\name{process_methylation_data}
\alias{process_methylation_data}
\title{Process methylation data using minfi pipeline. This function will drop samples with GT 5 percent high detection p-values, probes with at least one high detection p-value, Sex chromosomes, loci associated with single-nucleotide polymorphisms and cross-hybridizing probes. Additionally, will preprocess using Noob for background correction and dye bias normalization, BMIQ for probe design bias, batch correction using ComBat.}
\usage{
process_methylation_data(basenames)
}
\arguments{
\item{basenames}{list of idat file locations}
}
\value{
returns a methyl set
}
\description{
Process methylation data using minfi pipeline. This function will drop samples with GT 5 percent high detection p-values, probes with at least one high detection p-value, Sex chromosomes, loci associated with single-nucleotide polymorphisms and cross-hybridizing probes. Additionally, will preprocess using Noob for background correction and dye bias normalization, BMIQ for probe design bias, batch correction using ComBat.
}
