% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgcomp_wrap.R
\name{qgcomp_func}
\alias{qgcomp_func}
\title{Quantile g-computation wrapper function to automate analysis for multiple outcomes. Additionally, allows flexibility to perform stratified analysis.}
\usage{
qgcomp_func(
  outcomes,
  data,
  output_folder,
  chemicals,
  covariates,
  include_sex = TRUE,
  include_cohort = TRUE,
  q,
  b,
  seed
)
}
\arguments{
\item{outcomes}{input a list of outcome variables}

\item{output_folder}{complete location to a folder to save the output RDA files}

\item{chemicals}{list of variables that need to be considered as exposure mixture}

\item{covariates}{list of covariate variables}

\item{include_sex}{input TRUE or FALSE for sex specific stratification}

\item{include_cohort}{input TRUE or FALSE for cohort specific stratification}

\item{q}{input a numeric value for number of quantiles}

\item{b}{input a numeric value for number of random bootstrap samples}

\item{seed}{input a numeric value for random sample generation for bootstrap process}

\item{dat}{input dataset in tibble format}
}
\value{
This function will return RDA output files for each model run by boot and non-bootstrap method. Filenames will specify boot or nonboot, sex or cohort specific, and outcome name.
}
\description{
Quantile g-computation wrapper function to automate analysis for multiple outcomes. Additionally, allows flexibility to perform stratified analysis.
}
