% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwas_mummichog_PEA.R
\name{runMummichog}
\alias{runMummichog}
\title{Performs Pathway Enrichment Analysis (PEA) using mummichog function available from metaboAnalystR library
This function also prints the p-value theshold on scree that is being used for the analysis}
\usage{
runMummichog(
  input_folder,
  output_folder,
  p_val_cutoff = NULL,
  pick_500 = FALSE
)
}
\arguments{
\item{input_folder}{Specify folder location that contain the .txt files ready for the analysis. they should contain m.z, r.t, p.value variables}

\item{output_folder}{Specify location to save the output files}

\item{p_val_cutoff}{Specify a custom p-value or leave blank if "pick_500" is set to TRUE}

\item{pick_500}{Set this parameter to TRUE if you need the function to pick a p-value threshold based on 500th smallest p-value}
}
\value{
returns mummichog results in csv format at the output_folder location
}
\description{
Performs Pathway Enrichment Analysis (PEA) using mummichog function available from metaboAnalystR library
This function also prints the p-value theshold on scree that is being used for the analysis
}
\examples{
\donttest{
runMummichog(input_folder= "~/Documents/phth_phe_MWAS/result/enrich_exp",
            # p_val_cutoff= 0.2,
            pick_500 = TRUE,
            output_folder= "~/Documents/phth_phe_MWAS/result/enrich_imp")

}



}
