% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwas_split_lm_results_for_mummichog.R
\name{split_and_export}
\alias{split_and_export}
\title{Split MWAS LM result files from a single dataframe to .txt files for PEA mummichog}
\usage{
split_and_export(data, variables, export_location)
}
\arguments{
\item{data}{dataframe name that contains all the MWAS LM results combined}

\item{variables}{specify exposure variable as variable and mom/baby variable as set}

\item{export_location}{specify the folder location to save all the .txt files ready for mummichog analysis}
}
\value{
returns .txt files at the specified export_location folder
}
\description{
Split MWAS LM result files from a single dataframe to .txt files for PEA mummichog
}
\examples{
\donttest{
split_and_export(data= dat,
                variables = c("variable", "set"),
                export_location = "~/Documents/phth_phe_MWAS/result/enrich_exp/")

}

}
