% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{summarize_variables}
\alias{summarize_variables}
\title{Summarize categorical and continuous variables}
\usage{
summarize_variables(data, continuous_variables, categorical_variables)
}
\arguments{
\item{data}{input a dataframe that contain variables}

\item{continuous_variables}{provide list of categorical variables}

\item{categorical_variables}{provide list of continuous variables}
}
\value{
this function will return an object with 2 lists. one contain summary of continuous and other with categorical variables
}
\description{
Summarize categorical and continuous variables
}
