% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{var_selec}
\alias{var_selec}
\title{Variable selection using group LASSO (gglasso, grpreg that use distinct mathematical approaches), sparse group LASSO, glmnet, stability selection with error control}
\usage{
var_selec(X, Y, group)
}
\arguments{
\item{X}{input a matrix of exposure variables for selection}

\item{Y}{input a matrix of outcome variables}

\item{group}{input the grouping structure. Group inputs should be integers}
}
\value{
Returns a list of lists that contain beta coefficient for each exposure. Each list correspond to a outcome variable provided in the input
}
\description{
Variable selection using group LASSO (gglasso, grpreg that use distinct mathematical approaches), sparse group LASSO, glmnet, stability selection with error control
}
